refactor(product): Product 패키지 분리
