+++ src/main/java/kitchenpos/application/product/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.application.product.domain;
 
 import java.math.BigDecimal;
 import java.util.Objects;
