+++ src/main/java/kitchenpos/application/product/domain/ProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.domain.repository;
+package kitchenpos.application.product.domain;
 
-import kitchenpos.domain.Product;
+import kitchenpos.application.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
