+++ src/main/java/kitchenpos/application/product/service/ProductMapper.java
@@ -1,11 +1,11 @@
-package kitchenpos.application.mapper;
+package kitchenpos.application.product.service;
 
-import kitchenpos.application.dto.request.ProductRequest;
-import kitchenpos.domain.Product;
+import kitchenpos.application.product.domain.Product;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ProductMapper {
+public class
+ProductMapper {
     public Product mapFrom(ProductRequest productRequest) {
         return new Product(
                 productRequest.getName(),
