+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -10,15 +10,16 @@ import java.util.List;
 import kitchenpos.SpringBootTestWithProfiles;
 import kitchenpos.application.dto.request.MenuProductRequest;
 import kitchenpos.application.dto.request.MenuRequest;
-import kitchenpos.application.dto.request.ProductRequest;
+import kitchenpos.application.product.service.ProductRequest;
 import kitchenpos.application.dto.response.MenuGroupResponse;
 import kitchenpos.application.dto.response.MenuResponse;
-import kitchenpos.application.dto.response.ProductResponse;
+import kitchenpos.application.product.service.ProductResponse;
+import kitchenpos.application.product.service.ProductService;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.repository.MenuGroupRepository;
 import kitchenpos.domain.repository.MenuProductRepository;
 import kitchenpos.domain.repository.MenuRepository;
-import kitchenpos.domain.repository.ProductRepository;
+import kitchenpos.application.product.domain.ProductRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
