+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -14,13 +14,14 @@ import kitchenpos.application.dto.request.MenuRequest;
 import kitchenpos.application.dto.request.OrderRequest;
 import kitchenpos.application.dto.request.OrderRequest.OrderLineItemRequest;
 import kitchenpos.application.dto.request.OrderStatusRequest;
-import kitchenpos.application.dto.request.ProductRequest;
+import kitchenpos.application.product.service.ProductRequest;
 import kitchenpos.application.dto.request.TableRequest;
 import kitchenpos.application.dto.response.MenuGroupResponse;
 import kitchenpos.application.dto.response.MenuResponse;
 import kitchenpos.application.dto.response.OrderResponse;
 import kitchenpos.application.dto.response.OrderTableResponse;
-import kitchenpos.application.dto.response.ProductResponse;
+import kitchenpos.application.product.service.ProductResponse;
+import kitchenpos.application.product.service.ProductService;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.repository.MenuGroupRepository;
@@ -29,7 +30,7 @@ import kitchenpos.domain.repository.MenuRepository;
 import kitchenpos.domain.repository.OrderLineItemRepository;
 import kitchenpos.domain.repository.OrderRepository;
 import kitchenpos.domain.repository.OrderTableRepository;
-import kitchenpos.domain.repository.ProductRepository;
+import kitchenpos.application.product.domain.ProductRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
