+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -7,9 +7,10 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.SpringBootTestWithProfiles;
-import kitchenpos.application.dto.request.ProductRequest;
-import kitchenpos.application.dto.response.ProductResponse;
-import kitchenpos.domain.repository.ProductRepository;
+import kitchenpos.application.product.service.ProductRequest;
+import kitchenpos.application.product.service.ProductResponse;
+import kitchenpos.application.product.service.ProductService;
+import kitchenpos.application.product.domain.ProductRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
